code refactorized to maintain clean architecture directives
