--- app/src/main/java/com/example/imdb_v2/ui/theme/Color.kt
@@ -1,17 +1,0 @@
-package com.example.imdb_v2.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val Purple80 = Color(0xFFD0BCFF)
-val PurpleGrey80 = Color(0xFFCCC2DC)
-val Pink80 = Color(0xFFEFB8C8)
-
-val Purple40 = Color(0xFF6650a4)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
-
-val mainYellow = Color(0xFFF6C700)
-val mainBlack = Color(0xFF000000)
-val lightGray = Color(0xFF9D9C9C)
-val darkGray = Color(0xFF4B4747)
-val mainWhite = Color(0xFFF5F5F5)
