+++ app/src/main/java/com/example/imdb_v2/view/ui/theme/TextSnippets.kt
@@ -1,0 +1,21 @@
+package com.example.imdb_v2.view.ui.theme
+
+import androidx.compose.ui.text.font.Font
+import androidx.compose.ui.text.font.FontFamily
+import com.example.imdb_v2.R
+
+
+val robotoFamily = FontFamily(
+    Font(R.font.robotoblack),
+    Font(R.font.robotoblackitalic),
+    Font(R.font.robotobold),
+    Font(R.font.robotobolditalic),
+    Font(R.font.robotoitalic),
+    Font(R.font.robotolight),
+    Font(R.font.robotolightitalic),
+    Font(R.font.robotomedium),
+    Font(R.font.robotomediumitalic),
+    Font(R.font.robotoregular),
+    Font(R.font.robotothin),
+    Font(R.font.robotothinitalic),
+)
